% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelStructure.R
\name{as.MxRAMModel.lavaan}
\alias{as.MxRAMModel.lavaan}
\title{Extract path structure from a lavaan object}
\usage{
as.MxRAMModel.lavaan(model, exogenous = TRUE, standardized = FALSE,
  ...)
}
\arguments{
\item{model}{a lavaan or blavaan object}

\item{exogenous}{Compute causal influences of exogenous variables? (default TRUE)}

\item{standardized}{Compute causal influences for the standardized model? (default FALSE)}

\item{...}{Not currently used.}
}
\value{
An MxRAMModel containing the same path structure as the provided lavaan model
}
\description{
Extracts the paths from a lavaan object and returns them as an MxRAMModel
}
\details{
This function is experimental, and may have bugs.
At present, it does not allow for constraints,
}
