% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelStructure.R
\name{as.MxRAMModel.MxModel}
\alias{as.MxRAMModel.MxModel}
\title{Extract path structure from an OpenMx model}
\usage{
as.MxRAMModel.MxModel(model, exogenous = TRUE, standardized = FALSE,
  ...)
}
\arguments{
\item{model}{an OpenMx model}

\item{exogenous}{Compute causal influences of exogenous variables? (default TRUE)}

\item{standardized}{Compute causal influences for the standardized model? (default FALSE)}
}
\value{
An MxRAMModel containing the same path structure as the provided lavaan model
}
\description{
Extracts the paths from an OpenMx object and returns them as an MxRAMModel
  There's really no reason to use this.  Like, why are you even bothering?
}
\details{
This function is experimental, and may have bugs.
At present, it does not allow for constraints,
}
