% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelStructure.R
\name{as.MxRAMModel}
\alias{as.MxRAMModel}
\title{as.MxRAMModel: Create an MxRAMModel from a lavaan or OpenMx model object}
\usage{
as.MxRAMModel(model, exogenous = TRUE, standardized = FALSE, ...)
}
\arguments{
\item{model}{a path modeling object (see details for supported types)}

\item{exogenous}{Include exogenous variables? (default TRUE)}

\item{standardized}{Transform all variables into standardized forms? (default FALSE)}
}
\value{
An MxRAMModel containing the same path structure as the original model
}
\description{
Transforms a model into an MxRAMModel
}
\details{
This function is experimental, and may have bugs.
At present, it does not handle constraints, groups, or pretty much anything else
  that's at all complicated.

Currently supported: OpenMx RAM models (easy!), lavaan and blavaan models
}
