% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC.R
\name{computeHypotheticalSEs}
\alias{computeHypotheticalSEs}
\title{Compute hypothetical standard errors from an mxModel}
\usage{
computeHypotheticalSEs(model, N = NA, quiet = TRUE, ...)
}
\arguments{
\item{model}{The model}

\item{N}{The number of participants modeled.  If NA (the default) and the model has data, the size of the data is used.}

\item{quiet}{If TRUE, do not throw warnings (default: FALSE)}

\item{...}{Additional parameters for mxSE}
}
\value{
computed SEs, or NULL if computation was not possible.
}
\description{
Creates standard errors for the model assuming that the
model is exactly correct and complete data is available.
}
\details{
Computes the model-implied covariance, mean (if available)
and thresholds (if available) for the model, fits the model using
those as the data set, and computes standard errors for the MIC accordingly.
}
