% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC.R
\name{computeObservedSEs}
\alias{computeObservedSEs}
\title{Compute observed standard errors from an mxModel}
\usage{
computeObservedSEs(model, quiet = FALSE, ...)
}
\arguments{
\item{model}{The model}

\item{quiet}{If TRUE, do not throw warnings. (default: FALSE)}

\item{...}{Additional parameters for mxSE}
}
\value{
computed SEs, or NULL if computation was not possible.
}
\description{
Simply computes SEs for all total effects using mxSE, throwing appropriate errors for MICr.
}
