% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelStructure.R
\name{extractModelStructure}
\alias{extractModelStructure}
\title{Extract Model Structure and create an MxRAMModel}
\usage{
extractModelStructure(model, exogenous = TRUE, standardized = FALSE,
  ...)
}
\arguments{
\item{model}{a path modeling object (see details for supported types)}

\item{exogenous}{Include exogenous variables? (default TRUE)}

\item{standardized}{Transform all variables into standardized forms? (default FALSE)}
}
\value{
An MxRAMModel containing the same path structure as the original model
}
\description{
Extract paths from a model and convert to MxRAMModel
}
\details{
This function is experimental, and may have bugs.

At present, it does not handle constraints, groups, or pretty much anything else
  that's at all complicated.

Really, this is just a helper to handle the fact that blavaan doesn't export its blavaan class.

Currently supported: OpenMx RAM models (easy!), lavaan and blavaan models
}
