% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICTable.R
\name{flattenMIC}
\alias{flattenMIC}
\title{flattenMIC - Create}
\usage{
flattenMIC(mic, model = NULL, cullTiny = TRUE, tinyCutoff = 1e-06,
  includeModel = FALSE)
}
\arguments{
\item{mic}{An MIC matrix as returned by \code{\link{MIC}}}

\item{model}{the (character) label of the model, if different from what's included in the MIC}

\item{cullTiny}{whether or not to remove very small values from the table (default TRUE)}

\item{tinyCutoff}{how small is "very small" for culling (default 1e-6)}

\item{includeModel}{whether or not to include a separate column for model name in the table (default FALSE)}
}
\value{
The MIC table flattened form of this MIC
}
\description{
Generates a Matrix of Implied Causation for a given model, following the
  rules in Brick and Bailey (submitted).
}
\details{
Creates a flat-table version of a single MIC.
}
