% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mic_df.R
\name{get_group_mic_df}
\alias{get_group_mic_df}
\title{Function to compute indexes around mic for each hierarchy.}
\usage{
get_group_mic_df(df, group_key, min_value, ...)
}
\arguments{
\item{df}{: Data frame object which is input of mic.}

\item{group_key}{: characters, which is key of hierarchy}

\item{min_value}{: Minimum number of records in each group.}

\item{...}{: parameters for minerva::mine.}
}
\description{
Function to compute indexes around mic for each hierarchy.
}
