% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\docType{data}
\name{Images}
\alias{Images}
\title{Exemplary images}
\format{.png images.}
\source{
Some complicated machine.
}
\description{
Images bla bla
}
\examples{
# read all images
library(EBImage)
readImage(system.file('images/Well_Slide2_9_APC.png', package='micR'))
readImage(system.file('images/Well_Slide2_9_Cy3.png', package='micR'))
readImage(system.file('images/Well_Slide2_9_DAPI.png', package='micR'))
readImage(system.file('images/Well_Slide2_9_FITC.png', package='micR'))
}

