% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.dim.R
\name{img.dim}
\alias{img.dim}
\title{Determine a region of interest or preview area of an image object}
\usage{
img.dim(img.raw, width = 25, hight = 25)
}
\arguments{
\item{img.raw}{is the raw image that will used for the analysis}

\item{width}{is the width of the image area}

\item{hight}{is the hight of the image area}
}
\description{
Determine a region of interest or preview area of an image object
}
\details{
This function determines a region of interest or preview area of an 
image object. The imgtoEBImage function is used to read in the images.
}
\examples{
img.dim(system.file('images/Well_Slide2_9_APC.png', package='micR'), width = 25, hight = 25)
}
\author{
Stefan Rödiger
}
\references{
T.B.D.
}
\keyword{area}
\keyword{dimensions}

