% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.processor.R
\name{img.processor}
\alias{img.processor}
\title{Function to define a region of interest of a cell nucleus}
\usage{
img.processor(img.raw = NULL, gblur.sigma = 2, thresh.w = 20,
  thresh.h = 20, thresh.offset = 0.02, watershed.ext = 1)
}
\arguments{
\item{img.raw}{is the raw image that will used for the detection of the nucleus}

\item{gblur.sigma}{is the standard deviation of the Gaussian filter used for blurring (see gblur {EBImage})}

\item{thresh.w}{is the width of the moving rectangular window (see thresh {EBImage})}

\item{thresh.h}{is the height of the moving rectangular window (see thresh {EBImage})}

\item{thresh.offset}{is the thresholding offset from the averaged value (see thresh {EBImage})}

\item{watershed.ext}{is the radius of the neighborhood in pixels for the detection of neighboring 
objects. Higher value smoothes out small objects (see thresh {EBImage})}
}
\description{
Find nucleus of a cell.
}
\details{
This function is a pipeline of EBImage function to process an image 
with the aim to find region of interest from nuclei.
}
\author{
Stefan Roediger
}
\references{
Pau G, Fuchs F, Sklyar O, Boutros M and Huber W (2010). EBImage 
--an R package for image processing with applications to cellular phenotypes. 
Bioinformatics, 26(7), pp. 979--981. http://doi.org/10.1093/bioinformatics/btq046.
}
\keyword{blur}
\keyword{nucleus}
\keyword{watershed}

