% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.selector.R
\name{img.selector}
\alias{img.selector}
\title{Determines the object number of an EBImage processed image}
\usage{
img.selector(img.moment, min.diameter = 5, max.diameter = 55,
  pixel.length = 325)
}
\arguments{
\item{img.moment}{is the raw image that will used for the detection of the nucleus}

\item{min.diameter}{is the minimum diameter of a nucleus}

\item{max.diameter}{is the maximum diameter of a nucleus}

\item{pixel.length}{is the pixel length of the CCD chip (e.g., 325 nm)}
}
\description{
Selects objects based on a specific parameter value (currently nucleus diamter, only).
}
\details{
This function which image object are within a defined size range (e.g., betwenn 5 and 25 micro meter in diamter).
}
\author{
Stefan Rödiger
}
\references{
T.B.D.
}
\keyword{ROI}
\keyword{diameter}
\keyword{select}

