% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgtoEBImage.R
\name{imgtoEBImage}
\alias{imgtoEBImage}
\title{Helper function to read image files into EBImage}
\usage{
imgtoEBImage(file, convert.bmp = TRUE)
}
\arguments{
\item{file}{is the raw image (currently supported: bmp, tiff, png, jpeg).}

\item{convert.bmp}{is a binary operator used to convert the bmp integer matrix to floating point.}
}
\description{
Selects objects based on a specific moment parameter, where currently only majoraxis (recalculated as micro meter) is used.
}
\details{
This is a helper function to prepare pixel graphics for further use in EBImage.
}
\examples{
imgtoEBImage(system.file('images/Well_Slide2_9_APC.png', package='micR'))
}
\author{
Stefan Roediger
}
\references{
T.B.D.
}
\keyword{bmp}
\keyword{jpeg}
\keyword{png}
\keyword{tiff}

