% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micR-package.R
\docType{package}
\name{micR-package}
\alias{micR}
\alias{micR-package}
\title{Image processing with micR}
\description{
The package processes image data.
}
\author{
Stefan Rödiger.
Maintainer: Stefan Rödiger <Stefan.Roediger@b-tu.de>
}
\references{
T.B.D.
}
\keyword{package}

