% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC.R
\name{single.na}
\alias{single.na}
\title{single.na}
\usage{
single.na(a)
}
\arguments{
\item{a}{a thing}
}
\value{
Whether or not this is a single NA
}
\description{
is.na that appropriately handles things that are not single NAs
}
