% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelStructure.R
\name{standardizeMxRAMModel}
\alias{standardizeMxRAMModel}
\title{standardizeMxRAMModel: Create a new, standardized MxRAMModel from an old one.}
\usage{
standardizeMxRAMModel(model)
}
\arguments{
\item{model}{an MxRAMModel}
}
\value{
An MxRAMModel containing the same path structure as the original model, but standardized scores
}
\description{
Transforms an MxRAMModel into a Standardized MxRAMModel
}
\details{
This function is experimental, and may have bugs.
At present, it does not handle constraints, groups, or pretty much anything else
  that's at all complicated.
}
