% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelStructure.R
\name{as.MxRAMModel.character}
\alias{as.MxRAMModel.character}
\title{Extract path structure from a lavaan model string}
\usage{
as.MxRAMModel.character(model, exogenous = TRUE, standardized = FALSE,
  ...)
}
\arguments{
\item{model}{a lavaan model string}

\item{exogenous}{Compute causal influences of exogenous variables? (default TRUE)}

\item{standardized}{Compute causal influences for the standardized model? (default FALSE)}

\item{...}{other arguments to umxLav2RAM()}
}
\value{
An MxRAMModel containing the same path structure as the provided lavaan model
}
\description{
Extracts the paths from a lavaan syntax string and returns them as an MxRAMModel
}
\details{
This function is experimental, and may have bugs.
Under the hood, this function uses the umx function umxLav2RAM().  See the help file for that function for additional help.
}
