% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mic_df.R
\name{get_mic}
\alias{get_mic}
\title{Utility function to call get_mic_df and get_group_mic_df}
\usage{
get_mic(df, group_key = NA)
}
\arguments{
\item{df}{: Data frame object which is input of mic.}

\item{group_key}{: If NA, this function compute mic of flat df, otherwise mic of each group.}
}
\description{
Utility function to call get_mic_df and get_group_mic_df
}
