% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgtoEBImage.R
\name{imgtoEBImage}
\alias{imgtoEBImage}
\title{Helper function to read image files into EBImage}
\usage{
imgtoEBImage(file)
}
\arguments{
\item{file}{is the raw image (bmp).}
}
\description{
Selects objects based on a specific moment parameter, where currently only majoraxis (recalculated as micro meter) is used.
}
\details{
This is a helper function to prepare pixel graphics for further use in EBImage.
}
\author{
Stefan Roediger
}
\references{
T.B.D.
}
\keyword{bmp}

