% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC.R
\name{MIC}
\alias{MIC}
\title{MIC - Compute a matrix of implied causality}
\usage{
MIC(model, latents = TRUE, ...)
}
\arguments{
\item{model}{An Mx model in the RAM formulation}

\item{latents}{Compute causal influences of/on latent variables? (default TRUE)}

\item{...}{Other parameters passed to mxEval (e.g. defvar.row=)}
}
\value{
The MIC for this model.
}
\description{
Generates a Matrix of Implied Causation for a given model, following the
  rules in Brick and Bailey (submitted).
}
\details{
MICs trace all paths from all variables to all others, and present the
result as an asymmetric matrix.  The value of a single path is the product
of all single-headed arrows taken as part of that path, and the total causal
influence is the sum of all paths that connect one variable to another by
following only single-headed arrows, and only in the direction of the arrow.
}
\seealso{
MICTable
}
