% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micR-package.R
\docType{package}
\name{micR-package}
\alias{micR}
\alias{micR-package}
\title{Image processing with micR}
\description{
The package processes image data. micR combines several commonly required
functionality to import different images types, as supported by the R
environment. Functionality of packages from Bioconductor and CRAN are  
combined.
}
\author{
Stefan Roediger.
Maintainer: Stefan Roediger <Stefan.Roediger@b-tu.de>
}
\references{
T.B.D.
}
\keyword{package}

