\name{micR-package}
\alias{micR-package}
\alias{micR}
\docType{package}
\title{
\packageTitle{micR}
}
\description{
\packageDescription{micR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{micR}
\packageIndices{micR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{micR}

Maintainer: \packageMaintainer{micR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
