% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICTable.R
\name{MICTable}
\alias{MICTable}
\title{MICTable}
\usage{
MICTable(..., minAbs = 0.01, minDiff = NA, splitByType = FALSE)
}
\arguments{
\item{...}{MxModels from which to extract MICs}

\item{minAbs}{the smallest absolute size of causal effect that is worth noticing (default .01)}

\item{minDiff}{the smallest difference in effect that is worth considering a difference of scale (default minAbs)}

\item{splitByType}{If TRUE, return a list of tables split by the type of difference (e.g. existence of causal pathway, sign of causal path, scale of path, none of the above).  If FALSE, return a single table. Default TRUE.}
}
\value{
Three data frames:
 $exist contains causal effects that exist (|>minAbs|) in at least one model but do not exist in at least one other
 $sign contains causal effects that exist in all models, but with different signs in at least two models
 $scale contains causal effects with the same sign in all models, but which differ in scales (diff > minDiff) between at least two models
 $other causal effects that are identical between models
}
\description{
Takes a set of models and compares them all, following the rules in Brick and Bailey (2019).
}
\details{
Returns a comparison table comparing MICs from several models.
}
