% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spott.R
\name{spott}
\alias{spott}
\title{Calculate the central tendency of an image object}
\usage{
spott(img.raw, img.pp, img.moment, quantile = 0.03, ...)
}
\arguments{
\item{img.raw}{is the raw image that will used for the detection of the nucleus}

\item{img.pp}{is input of the type img.processor object (see img.processor {micR})}

\item{img.moment}{is an computeFeatures.moment object (see computeFeatures.moment {EBImage})}

\item{quantile}{is a value to define the values removed by quantile filtering}

\item{...}{is the input for the img.selector subroutine (see img.selector {micR})}
}
\description{
Selects objects based on a specific moment parameter, where currently only majoraxis (recalculated as micro meter) is used.
}
\details{
This function calculates the mean of a quantile filtered image object.
}
\author{
Stefan Roediger
}
\references{
T.B.D.
}
\keyword{mean}
\keyword{quantile}

